Hop along multiple inputs simultaneously

hop2() and phop() represent the combination of slide2() and pslide() with hop(), allowing you to iterate over multiple vectors at once, hopping along them using boundaries defined by .starts and .stops.

hop2(.x, .y, .starts, .stops, .f, ...)

hop2_vec(.x, .y, .starts, .stops, .f, ..., .ptype = list())

hop2_dbl(.x, .y, .starts, .stops, .f, ...)

hop2_int(.x, .y, .starts, .stops, .f, ...)

hop2_lgl(.x, .y, .starts, .stops, .f, ...)

hop2_chr(.x, .y, .starts, .stops, .f, ...)

hop2_raw(.x, .y, .starts, .stops, .f, ...)

hop2_dfr(
  .x,
  .y,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

hop2_dfc(
  .x,
  .y,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)

phop(.l, .starts, .stops, .f, ...)

phop_vec(.l, .starts, .stops, .f, ..., .ptype = list())

phop_dbl(.l, .starts, .stops, .f, ...)

phop_int(.l, .starts, .stops, .f, ...)

phop_lgl(.l, .starts, .stops, .f, ...)

phop_chr(.l, .starts, .stops, .f, ...)

phop_raw(.l, .starts, .stops, .f, ...)

phop_dfr(
  .l,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

phop_dfc(
  .l,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)

Arguments

.x, .y
[vector]

Vectors to iterate over. Vectors of size 1 will be recycled.

.starts
[vector]

Vectors of boundary locations that make up the windows to bucket .x with. Both .starts and .stops will be recycled to their common size, and that common size will be the size of the result. Both vectors should be integer locations along .x.

.stops
[vector]

Vectors of boundary locations that make up the windows to bucket .x with. Both .starts and .stops will be recycled to their common size, and that common size will be the size of the result. Both vectors should be integer locations along .x.

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.ptype
[vector]

The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

.l
[list]

A list of vectors. The length of .l determines the number of arguments that .f will be called with. If .l has names, they will be used as named arguments to .f. Elements of .l with size 1 will be recycled.

Invariants

hop2()

vec_size(hop2(.x, .y, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_ptype(hop2(.x, .y, .starts, .stops)) == list()

hop2_vec() and hop2_*() variants

vec_size(hop2_vec(.x, .y, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_size(hop2_vec(.x, .y, .starts, .stops)[[1]]) == 1L

vec_ptype(hop2_vec(.x, .y, .starts, .stops, .ptype = ptype)) == ptype

phop()

vec_size(phop(.l, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_ptype(phop(.l, .starts, .stops)) == list()

phop_vec() and phop_*() variants

vec_size(phop_vec(.l, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_size(phop_vec(.l, .starts, .stops)[[1]]) == 1L

vec_ptype(phop_vec(.l, .starts, .stops, .ptype = ptype)) == ptype

See also

Examples

hop2(1:2, 3:4, .starts = 1, .stops = c(2, 1), ~c(x = .x, y = .y))

#> [[1]]
#> x1 x2 y1 y2 
#>  1  2  3  4 
#> 
#> [[2]]
#> x y 
#> 1 3 
#> 

phop(
 list(1, 2:4, 5:7),
 .starts = c(0, 1),
 .stops  = c(2, 4),
 ~c(x = ..1, y = ..2, z = ..3)
)

#> [[1]]
#> x1 x2 y1 y2 z1 z2 
#>  1  1  2  3  5  6 
#> 
#> [[2]]
#> x1 x2 x3 y1 y2 y3 z1 z2 z3 
#>  1  1  1  2  3  4  5  6  7 
#>