Slide over multiple inputs simultaneously

These are variants of slide() that iterate over multiple inputs in parallel. They are parallel in the sense that each input is processed in parallel with the others, not in the sense of multicore computing. These functions work similarly to map2() and pmap() from purrr.

slide2(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_vec(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE,
  .ptype = list())

slide2_dbl(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_int(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_lgl(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_chr(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_raw(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

slide2_dfr(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE,
  .names_to = NULL, .name_repair = c("unique", "universal",
  "check_unique"))

slide2_dfc(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE, .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal"))

pslide(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_vec(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE,
  .ptype = list())

pslide_dbl(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_int(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_lgl(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_chr(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_raw(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE)

pslide_dfr(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE,
  .names_to = NULL, .name_repair = c("unique", "universal",
  "check_unique"))

pslide_dfc(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .complete = FALSE, .forward = TRUE, .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal"))

Arguments

.x, .y
[vector] Vectors to iterate over. Vectors of size 1 will be recycled.

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.before
[integer] The number of values before the current element to include in the sliding window. Set to unbounded() to select all elements before the current position. A negative value is allowed, and allows you to "look forward" as well.

.after
[integer] The number of values after the current element to include in the sliding window. Set to unbounded() to select all elements after the current position. A negative value is allowed, and allows you to "look backward" as well.

.step
[positive integer] The number of elements to shift the window forward (or backward, depending on .forward) between function calls.

.offset
[NULL / positive integer] An offset from the beginning (or end, depending on .forward) of .x to place the first element in the output vector. If NULL, this is computed automatically as the first location where a complete sliding window can be generated.

.complete
[logical] Should the sliding be restricted to complete windows only? If FALSE, the default, then partial computations will be allowed.

.forward
[logical] The direction to slide.

.ptype
[vector] The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

.l
[list] A list of vectors. The length of .l determines the number of arguments that .f will be called with. If .l has names, they will be used as named arguments to .f.

Examples

# Slide along two inputs at once
slide2(1:5, 6:10, ~list(.x, .y), .before = 2)

#> [[1]]
#> [[1]][[1]]
#> [1] 1
#> 
#> [[1]][[2]]
#> [1] 6
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 1 2
#> 
#> [[2]][[2]]
#> [1] 6 7
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] 1 2 3
#> 
#> [[3]][[2]]
#> [1] 6 7 8
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] 2 3 4
#> 
#> [[4]][[2]]
#> [1] 7 8 9
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#> [1] 3 4 5
#> 
#> [[5]][[2]]
#> [1]  8  9 10
#> 
#> 

# Or, for more than two, use `pslide()`
pslide(list(1:5, 6:10, 11:15), ~list(.x, .y, ..3), .before = 2)

#> [[1]]
#> [[1]][[1]]
#> [1] 1
#> 
#> [[1]][[2]]
#> [1] 6
#> 
#> [[1]][[3]]
#> [1] 11
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 1 2
#> 
#> [[2]][[2]]
#> [1] 6 7
#> 
#> [[2]][[3]]
#> [1] 11 12
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] 1 2 3
#> 
#> [[3]][[2]]
#> [1] 6 7 8
#> 
#> [[3]][[3]]
#> [1] 11 12 13
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] 2 3 4
#> 
#> [[4]][[2]]
#> [1] 7 8 9
#> 
#> [[4]][[3]]
#> [1] 12 13 14
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#> [1] 3 4 5
#> 
#> [[5]][[2]]
#> [1]  8  9 10
#> 
#> [[5]][[3]]
#> [1] 13 14 15
#> 
#> 

# You can even slide along the rows of multiple data frames of
# equal size at once
x <- data.frame(a = rnorm(10), b = rnorm(10))
y <- data.frame(c = letters[1:10], d = letters[11:20])

row_return <- function(x_rows, y_rows) {
  if (sum(x_rows$a) < 0) {
    x_rows
  } else {
    y_rows
  }
}

slide2(x, y, row_return, .before = 1, .after = 2)

#> [[1]]
#>            a         b
#> 1  0.4681544 0.9353632
#> 2  0.3629513 0.1764886
#> 3 -1.3045435 0.2436855
#> 
#> [[2]]
#>   c d
#> 1 a k
#> 2 b l
#> 3 c m
#> 4 d n
#> 
#> [[3]]
#>   c d
#> 1 b l
#> 2 c m
#> 3 d n
#> 4 e o
#> 
#> [[4]]
#>   c d
#> 1 c m
#> 2 d n
#> 3 e o
#> 4 f p
#> 
#> [[5]]
#>   c d
#> 1 d n
#> 2 e o
#> 3 f p
#> 4 g q
#> 
#> [[6]]
#>   c d
#> 1 e o
#> 2 f p
#> 3 g q
#> 4 h r
#> 
#> [[7]]
#>             a          b
#> 1 -0.09744510 -0.1339970
#> 2 -0.93584735 -1.9100875
#> 3 -0.01595031 -0.2792372
#> 4 -0.82678895 -0.3134460
#> 
#> [[8]]
#>             a          b
#> 1 -0.93584735 -1.9100875
#> 2 -0.01595031 -0.2792372
#> 3 -0.82678895 -0.3134460
#> 4 -1.51239965  1.0673079
#> 
#> [[9]]
#>             a          b
#> 1 -0.01595031 -0.2792372
#> 2 -0.82678895 -0.3134460
#> 3 -1.51239965  1.0673079
#> 
#> [[10]]
#>           a         b
#> 1 -0.826789 -0.313446
#> 2 -1.512400  1.067308
#>