Slide between boundaries

slide_between() is the lower level engine that powers slide_index(). It has slightly different invariants than slide_index(), and is useful when you either need to hand craft boundary values, or want to compute a result with a size that is different from .x.

slide_between(.x, .i, .starts, .stops, .f, ...)

slide_between_vec(.x, .i, .starts, .stops, .f, ..., .ptype = list())

slide_between_dbl(.x, .i, .starts, .stops, .f, ...)

slide_between_int(.x, .i, .starts, .stops, .f, ...)

slide_between_lgl(.x, .i, .starts, .stops, .f, ...)

slide_between_chr(.x, .i, .starts, .stops, .f, ...)

slide_between_raw(.x, .i, .starts, .stops, .f, ...)

slide_between_dfr(.x, .i, .starts, .stops, .f, ..., .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique"))

slide_between_dfc(.x, .i, .starts, .stops, .f, ..., .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal"))

Arguments

.x
[vector]

The vector to iterate over and apply .f to.

.i
[vector]

The index vector that determines the window sizes. The lower bound of the window range will be computed as .i - .before, and the upper bound will be .i + .after. It is faily common to supply a date vector as the index, but not required.

.starts, .stops
[vector]

Vectors of boundary values that make up the windows to bucket .i with. Both .starts and .stops will be recycled to their common size, and that common size will be the size of the result. Both vectors should be the same type as .i. These boundaries are both inclusive, meaning that .i will be searched for values that fall into the range of [start, stop].

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.ptype
[vector]

The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

Invariants

slide_between()

vec_size(slide_between(.x, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_ptype(slide_between(.x, .starts, .stops)) == list()

slide_between_vec() and slide_between_*() variants

vec_size(slide_between_vec(.x, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_size(slide_between_vec(.x, .starts, .stops)[[1]]) == 1L

vec_ptype(slide_between_vec(.x, .starts, .stops, .ptype = ptype)) == ptype

The .i-ndex

There are 3 restrictions on .i:

The size of .i must match the size of .x, .i will not be recycled.

.i must be an increasing vector, but duplicate values are allowed.

.i is not allowed to have missing values.

See also

Examples

library(vctrs)
library(lubridate, warn.conflicts = FALSE)

# ---------------------------------------------------------------------------
# Returning a size smaller than `.x`

i <- as.Date("2019-01-25") + c(0, 1, 2, 3, 10, 20, 35, 42, 45)

# slide_index() allows you to slide relative to `i`
slide_index(i, i, ~.x, .before = weeks(1))

#> [[1]]
#> [1] "2019-01-25"
#> 
#> [[2]]
#> [1] "2019-01-25" "2019-01-26"
#> 
#> [[3]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27"
#> 
#> [[4]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[5]]
#> [1] "2019-01-28" "2019-02-04"
#> 
#> [[6]]
#> [1] "2019-02-14"
#> 
#> [[7]]
#> [1] "2019-03-01"
#> 
#> [[8]]
#> [1] "2019-03-01" "2019-03-08"
#> 
#> [[9]]
#> [1] "2019-03-08" "2019-03-11"
#> 

# But you might be more interested in coarser summaries. This groups
# by year-month and computes 2 `.f` on 2 month windows.
i_yearmonth <- year(i) + (month(i) - 1) / 12
slide_index(i, i_yearmonth, ~.x, .before = 1)

#> [[1]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[2]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[3]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[4]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[5]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28" "2019-02-04"
#> [6] "2019-02-14"
#> 
#> [[6]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28" "2019-02-04"
#> [6] "2019-02-14"
#> 
#> [[7]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28" "2019-02-04"
#> [6] "2019-02-14" "2019-03-01" "2019-03-08" "2019-03-11"
#> 
#> [[8]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28" "2019-02-04"
#> [6] "2019-02-14" "2019-03-01" "2019-03-08" "2019-03-11"
#> 
#> [[9]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28" "2019-02-04"
#> [6] "2019-02-14" "2019-03-01" "2019-03-08" "2019-03-11"
#> 

# ^ This works nicely when working with dplyr if you are trying to create
# a new column in a data frame, but you'll notice that there are really only
# 3 months, so only 3 values are being calculated. If you only want to return
# a vector of those 3 values, you can use `slide_between()`. You'll have to
# hand craft the boundaries, but this is a general strategy
# I've found useful:
first_start <- floor_date(i[1], "months")
last_stop <- ceiling_date(i[length(i)], "months")
dates <- seq(first_start, last_stop, "1 month")
inner <- dates[2:(length(dates) - 1L)]
starts <- vec_c(first_start, inner)
stops <- vec_c(inner - 1, last_stop)

slide_between(i, i, starts, stops, ~.x)

#> [[1]]
#> [1] "2019-01-25" "2019-01-26" "2019-01-27" "2019-01-28"
#> 
#> [[2]]
#> [1] "2019-02-04" "2019-02-14"
#> 
#> [[3]]
#> [1] "2019-03-01" "2019-03-08" "2019-03-11"
#> 

# ---------------------------------------------------------------------------
# Non-existant dates with `lubridate::months()`

# Imagine you want to compute a 1 month rolling average on this
# irregular daily data.
i <- vec_c(as.Date("2019-02-27") + 0:3, as.Date("2019-03-27") + 0:5)
x <- rnorm(vec_seq_along(i))

# You might try `slide_index()` like this, but you'd run into this error
cnd <- try(slide_index(x, i, mean, .before = months(1)), silent = TRUE)
attr(cnd, "condition")[["message"]]

#> The range generated by `.before` cannot have `NA` values, which were found at location(s): 7, 8, 9.

# This is because when you actually compute the `.i - .before` sequence,
# you hit non-existant dates. i.e. `"2019-03-29" - months(1)` doesn't exist.
i - months(1)

#>  [1] "2019-01-27" "2019-01-28" "2019-02-01" "2019-02-02" "2019-02-27"
#>  [6] "2019-02-28" NA           NA           NA           "2019-03-01"

# To get around this, lubridate provides `add_with_rollback()`,
# and the shortcut operation `%m-%`, which subtracts the month, then rolls
# forward/backward if it hits an `NA`. You can manually generate boundaries,
# then provide them to `slide_between()`.
starts <- i %m-% months(1)
stops <- i

slide_between(x, i, starts, stops, mean)

#> [[1]]
#> [1] 1.847182
#> 
#> [[2]]
#> [1] 0.9795577
#> 
#> [[3]]
#> [1] 0.4043594
#> 
#> [[4]]
#> [1] 0.717823
#> 
#> [[5]]
#> [1] 0.7186025
#> 
#> [[6]]
#> [1] 0.01654996
#> 
#> [[7]]
#> [1] 0.1097766
#> 
#> [[8]]
#> [1] 0.1616313
#> 
#> [[9]]
#> [1] 0.07358597
#> 
#> [[10]]
#> [1] 0.2005552
#> 

slide_between(i, i, starts, stops, ~.x)

#> [[1]]
#> [1] "2019-02-27"
#> 
#> [[2]]
#> [1] "2019-02-27" "2019-02-28"
#> 
#> [[3]]
#> [1] "2019-02-27" "2019-02-28" "2019-03-01"
#> 
#> [[4]]
#> [1] "2019-02-27" "2019-02-28" "2019-03-01" "2019-03-02"
#> 
#> [[5]]
#> [1] "2019-02-27" "2019-02-28" "2019-03-01" "2019-03-02" "2019-03-27"
#> 
#> [[6]]
#> [1] "2019-02-28" "2019-03-01" "2019-03-02" "2019-03-27" "2019-03-28"
#> 
#> [[7]]
#> [1] "2019-02-28" "2019-03-01" "2019-03-02" "2019-03-27" "2019-03-28"
#> [6] "2019-03-29"
#> 
#> [[8]]
#> [1] "2019-02-28" "2019-03-01" "2019-03-02" "2019-03-27" "2019-03-28"
#> [6] "2019-03-29" "2019-03-30"
#> 
#> [[9]]
#> [1] "2019-02-28" "2019-03-01" "2019-03-02" "2019-03-27" "2019-03-28"
#> [6] "2019-03-29" "2019-03-30" "2019-03-31"
#> 
#> [[10]]
#> [1] "2019-03-01" "2019-03-02" "2019-03-27" "2019-03-28" "2019-03-29"
#> [6] "2019-03-30" "2019-03-31" "2019-04-01"
#> 

# ---------------------------------------------------------------------------
# Business calendars with RcppQuantuccia

# Imagine you work in a company with daily costs that fall on these dates.
# Notice that these are strictly weekdays that cross over a weekend.
i <- as.Date("2019-08-15") + c(0:1, 4, 6, 7)
wday(i, label = TRUE)

#> [1] Thu Fri Mon Wed Thu
#> Levels: Sun < Mon < Tue < Wed < Thu < Fri < Sat

# If you are following a "business calendar" and want to compute a rolling
# 2 day computation, you might want `"2019-08-19"` (a Monday) to be paired
# with `"2019-08-16"` (a Friday) because they are adjacent in the business
# week. Using the RcppQuantuccia package, we can use `advanceUnits()` to
# compute the correct business day boundaries for use in `slide_between()`.
library(RcppQuantuccia)
starts <- advanceUnits(i, -1, "Days")
stops <- i

# "The current day + 1 business day before"
slide_between(i, i, starts, stops, ~.x)

#> [[1]]
#> [1] "2019-08-15"
#> 
#> [[2]]
#> [1] "2019-08-15" "2019-08-16"
#> 
#> [[3]]
#> [1] "2019-08-16" "2019-08-19"
#> 
#> [[4]]
#> [1] "2019-08-21"
#> 
#> [[5]]
#> [1] "2019-08-21" "2019-08-22"
#>