Slide over multiple inputs between boundaries

slide_between2() and pslide_between() represent the combination of slide2() and pslide() with slide_between(), allowing you to iterate over multiple vectors at once, while sliding along an .i-ndex with boundaries defined by .starts and .stops.

slide_between2(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_vec(.x, .y, .i, .starts, .stops, .f, ..., .ptype = list())

slide_between2_dbl(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_int(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_lgl(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_chr(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_raw(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_dfr(
  .x,
  .y,
  .i,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

slide_between2_dfc(
  .x,
  .y,
  .i,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)

pslide_between(.l, .i, .starts, .stops, .f, ...)

pslide_between_vec(.l, .i, .starts, .stops, .f, ..., .ptype = list())

pslide_between_dbl(.l, .i, .starts, .stops, .f, ...)

pslide_between_int(.l, .i, .starts, .stops, .f, ...)

pslide_between_lgl(.l, .i, .starts, .stops, .f, ...)

pslide_between_chr(.l, .i, .starts, .stops, .f, ...)

pslide_between_raw(.l, .i, .starts, .stops, .f, ...)

pslide_between_dfr(
  .l,
  .i,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

pslide_between_dfc(
  .l,
  .i,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)

Arguments

.x, .y
[vector]

Vectors to iterate over. Vectors of size 1 will be recycled.

.i
[vector]

The index vector that determines the window sizes. The lower bound of the window range will be computed as .i - .before, and the upper bound will be .i + .after. It is faily common to supply a date vector as the index, but not required.

.starts
[vector]

Vectors of boundary values that make up the windows to bucket .i with. Both .starts and .stops will be recycled to their common size, and that common size will be the size of the result. Both vectors should be the same type as .i. These boundaries are both inclusive, meaning that .i will be searched for values that fall into the range of [start, stop].

.stops
[vector]

Vectors of boundary values that make up the windows to bucket .i with. Both .starts and .stops will be recycled to their common size, and that common size will be the size of the result. Both vectors should be the same type as .i. These boundaries are both inclusive, meaning that .i will be searched for values that fall into the range of [start, stop].

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.ptype
[vector]

The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

.l
[list]

A list of vectors. The length of .l determines the number of arguments that .f will be called with. If .l has names, they will be used as named arguments to .f. Elements of .l with size 1 will be recycled.

Invariants

slide_between2()

vec_size(slide_between2(.x, .y, .starts, .stops)) == vvec_size_common(.starts, .stops)

vec_ptype(slide_between2(.x, .y, .starts, .stops)) == list()

slide_between2_vec() and slide_between2_*() variants

vec_size(slide_between2_vec(.x, .y, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_size(slide_between2_vec(.x, .y, .starts, .stops)[[1]]) == 1L

vec_ptype(slide_between2_vec(.x, .y, .starts, .stops, .ptype = ptype)) == ptype

pslide_between()

vec_size(pslide_between(.l, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_ptype(pslide_between(.l, .starts, .stops)) == list()

pslide_between_vec() and pslide_between_*() variants

vec_size(pslide_between_vec(.l, .starts, .stops)) == vec_size_common(.starts, .stops)

vec_size(pslide_between_vec(.l, .starts, .stops)[[1]]) == 1L

vec_ptype(pslide_between_vec(.l, .starts, .stops, .ptype = ptype)) == ptype

The .i-ndex

There are 3 restrictions on .i:

The size of .i must match the size of .x, .i will not be recycled.

.i must be an increasing vector, but duplicate values are allowed.

.i is not allowed to have missing values.

See also

Examples

# Notice that `i` is an irregular index!
x <- 1:5
i <- as.Date("2019-08-15") + c(0:1, 4, 6, 7)

# Manually create starts/stops. They don't have to be equally spaced,
# and they don't have to be the same size as `.x` or `.i`.
starts <- as.Date(c("2019-08-15", "2019-08-18"))
stops <- as.Date(c("2019-08-16", "2019-08-23"))

# The output size is equal to the common size of `.starts` and `.stops`
slide_between2(x, i, i, starts, stops, ~data.frame(x = .x, y = .y))

#> [[1]]
#>   x          y
#> 1 1 2019-08-15
#> 2 2 2019-08-16
#> 
#> [[2]]
#>   x          y
#> 1 3 2019-08-19
#> 2 4 2019-08-21
#> 3 5 2019-08-22
#>