Slide along multiples inputs simultaneously over an index

slide_index2() and pslide_index() represent the combination of slide2() and pslide() with slide_index(), allowing you to iterate over multiple vectors at once, while respecting an .i-ndex.

slide_index2(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_vec(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .ptype = list())

slide_index2_dbl(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_int(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_lgl(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_chr(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_raw(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

slide_index2_dfr(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .names_to = NULL, .name_repair = c("unique",
  "universal", "check_unique"))

slide_index2_dfc(.x, .y, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .size = NULL, .name_repair = c("unique",
  "universal", "check_unique", "minimal"))

pslide_index(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_vec(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .ptype = list())

pslide_index_dbl(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_int(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_lgl(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_chr(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_raw(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE)

pslide_index_dfr(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .names_to = NULL, .name_repair = c("unique",
  "universal", "check_unique"))

pslide_index_dfc(.l, .i, .f, ..., .before = 0L, .after = 0L,
  .complete = FALSE, .size = NULL, .name_repair = c("unique",
  "universal", "check_unique", "minimal"))

Arguments

.x, .y
[vector]

Vectors to iterate over. Vectors of size 1 will be recycled.

.i
[vector]

The index vector that determines the window sizes. The lower bound of the window range will be computed as .i - .before, and the upper bound will be .i + .after. It is faily common to supply a date vector as the index, but not required.

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.before
[vector(1) / function / formula]

The number of values before the current element of .i to include in the sliding window. Set to unbounded() to select all elements before the current element, constructing a cumulative window. A negative value is allowed, and allows you to "look forward" as well.

For slide_index(), this can be any object that can be subtracted from .i with -. One common use case is to set this to a lubridate period, such as lubridate::weeks().

Additionally, .before is allowed to be a function or lambda function with 1 argument, the entire .i vector.

.after
[vector(1) / function / formula]

The number of values after the current element of .i to include in the sliding window. Set to unbounded() to select all elements after the current position, constructing the reverse of a cumulative window, where you start with as many elements as possible and decrease the amount as you move through .x. A negative value is allowed, and allows you to "look backward" as well.

For slide_index(), this can be any object that can be added to .i with +. One common use case is to set this to a lubridate period, such as lubridate::weeks().

Additionally, .after is allowed to be a function or lambda function with 1 argument, the entire .i vector.

.complete
[logical(1)]

Should the sliding be restricted to complete windows only? If FALSE, the default, then partial computations will be allowed.

.ptype
[vector]

The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

.l
[list]

A list of vectors. The length of .l determines the number of arguments that .f will be called with. If .l has names, they will be used as named arguments to .f. Elements of .l with size 1 will be recycled.

.i Details

There are 3 restrictions on .i:

The size of .i must match the size of .x, .i will not be recycled.

.i must be an increasing vector, but duplicate values are allowed.

.i is not allowed to have missing values.

The ranges computed from .i - .before, .i + .after, or by applying a before / after function to .i have the same restrictions as the three mentioned above.

See also

Examples

# Notice that `i` is an irregular index!
x <- 1:5
y <- 6:10
i <- as.Date("2019-08-15") + c(0:1, 4, 6, 7)

# When we slide over `i` looking back 1 day, the irregularity is respected.
# When there is a gap in dates, only 2 values are returned (one from
# `x` and one from `y`), otherwise, 4 values are returned.
slide_index2(x, y, i, ~c(.x, .y), .before = 1)

#> [[1]]
#> [1] 1 6
#> 
#> [[2]]
#> [1] 1 2 6 7
#> 
#> [[3]]
#> [1] 3 8
#> 
#> [[4]]
#> [1] 4 9
#> 
#> [[5]]
#> [1]  4  5  9 10
#>