Slide over multiple inputs simultaneously

These are variants of slide() that iterate over multiple inputs in parallel. They are parallel in the sense that each input is processed in parallel with the others, not in the sense of multicore computing. These functions work similarly to map2() and pmap() from purrr.

slide2(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_vec(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .ptype = list())

slide2_dbl(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_int(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_lgl(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_chr(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_raw(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

slide2_dfr(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .names_to = NULL, .name_repair = c("unique",
  "universal", "check_unique"))

slide2_dfc(.x, .y, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .size = NULL, .name_repair = c("unique",
  "universal", "check_unique", "minimal"))

pslide(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_vec(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .ptype = list())

pslide_dbl(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_int(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_lgl(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_chr(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_raw(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE)

pslide_dfr(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .names_to = NULL, .name_repair = c("unique",
  "universal", "check_unique"))

pslide_dfc(.l, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .complete = FALSE, .size = NULL, .name_repair = c("unique",
  "universal", "check_unique", "minimal"))

Arguments

.x, .y
[vector]

Vectors to iterate over. Vectors of size 1 will be recycled.

.f
[function / formula]

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function with up to two arguments: .x (single argument) or .x and .y (two arguments). The . placeholder can be used instead of .x. This allows you to create very compact anonymous functions with up to two inputs.

...
Additional arguments passed on to the mapped function.

.before
[integer(1) / Inf]

The number of values before the current element to include in the sliding window. Set to Inf to select all elements before the current position, constructing a cumulative window. A negative value is allowed, and allows you to "look forward" as well.

.after
[integer(1) / Inf]

The number of values after the current element to include in the sliding window. Set to Inf to select all elements after the current position, constructing the reverse of a cumulative window, where you start with as many elements as possible and decrease the amount as you move through .x. A negative value is allowed, and allows you to "look backward" as well.

.step
[positive integer(1)]

The number of elements to shift the window forward between function calls.

.complete
[logical(1)]

Should the sliding be restricted to complete windows only? If FALSE, the default, then partial computations will be allowed.

.ptype
[vector]

The prototype corresponding to the type of the output. Defaults to a list().

.names_to
Optionally, the name of a column where the names of ... arguments are copied. These names are useful to identify which row comes from which input. If supplied and ... is not named, an integer column is used to identify the rows.

.name_repair
One of "unique", "universal", or "check_unique". See vec_as_names() for the meaning of these options.

With vec_rbind(), the repair function is applied to all inputs separately. This is because vec_rbind() needs to align their columns before binding the rows, and thus needs all inputs to have unique names. On the other hand, vec_cbind() applies the repair function after all inputs have been concatenated together in a final data frame. Hence vec_cbind() allows the more permissive minimal names repair.

.size
If, NULL, the default, will determine the number of rows in vec_cbind() output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs of length 1 will be recycled appropriately.

.l
[list]

A list of vectors. The length of .l determines the number of arguments that .f will be called with. If .l has names, they will be used as named arguments to .f. Elements of .l with size 1 will be recycled.

Invariants

slide2()

vec_size(slide2(.x, .y)) == vec_size_common(.x, .y)

vec_ptype(slide2(.x, .y)) == list()

slide2_vec() and slide2_*() variants

vec_size(slide2_vec(.x, .y)) == vec_size_common(.x, .y)

vec_size(slide2_vec(.x, .y)[[1]]) == 1L

vec_ptype(slide2_vec(.x, .y, .ptype = ptype)) == ptype

pslide()

vec_size(pslide(.l)) == vec_size_common(!!! .l)

vec_ptype(pslide(.l)) == list()

pslide_vec() and pslide_*() variants

vec_size(pslide_vec(.l)) == vec_size_common(!!! .l)

vec_size(pslide_vec(.l)[[1]]) == 1L

vec_ptype(pslide_vec(.l, .ptype = ptype)) == ptype

See also

Examples

# Slide along two inputs at once
slide2(1:4, 5:8, ~list(.x, .y), .before = 2)

#> [[1]]
#> [[1]][[1]]
#> [1] 1
#> 
#> [[1]][[2]]
#> [1] 5
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 1 2
#> 
#> [[2]][[2]]
#> [1] 5 6
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] 1 2 3
#> 
#> [[3]][[2]]
#> [1] 5 6 7
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] 2 3 4
#> 
#> [[4]][[2]]
#> [1] 6 7 8
#> 
#> 

# Or, for more than two, use `pslide()`
pslide(list(1:4, 5:8, 9:12), ~list(.x, .y, ..3), .before = 2)

#> [[1]]
#> [[1]][[1]]
#> [1] 1
#> 
#> [[1]][[2]]
#> [1] 5
#> 
#> [[1]][[3]]
#> [1] 9
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 1 2
#> 
#> [[2]][[2]]
#> [1] 5 6
#> 
#> [[2]][[3]]
#> [1]  9 10
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] 1 2 3
#> 
#> [[3]][[2]]
#> [1] 5 6 7
#> 
#> [[3]][[3]]
#> [1]  9 10 11
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] 2 3 4
#> 
#> [[4]][[2]]
#> [1] 6 7 8
#> 
#> [[4]][[3]]
#> [1] 10 11 12
#> 
#> 

# You can even slide along the rows of multiple data frames of
# equal size at once
set.seed(16)
x <- data.frame(a = rnorm(5), b = rnorm(5))
y <- data.frame(c = letters[1:5], d = letters[6:10])

row_return <- function(x_rows, y_rows) {
  if (sum(x_rows$a) < 0) {
    x_rows
  } else {
    y_rows
  }
}

slide2(x, y, row_return, .before = 1, .after = 2)

#> [[1]]
#>   c d
#> 1 a f
#> 2 b g
#> 3 c h
#> 
#> [[2]]
#>   c d
#> 1 a f
#> 2 b g
#> 3 c h
#> 4 d i
#> 
#> [[3]]
#>   c d
#> 1 b g
#> 2 c h
#> 3 d i
#> 4 e j
#> 
#> [[4]]
#>   c d
#> 1 c h
#> 2 d i
#> 3 e j
#> 
#> [[5]]
#>           a        b
#> 1 -1.444229 1.024973
#> 2  1.147829 0.573142
#>